pyproject.toml
[project]
name = "flyer-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.13",
    "langchain>=0.3.26",
    "langchain-core>=0.3.66",
    "langchain-openai>=0.3.25",
    "pillow>=11.2.1",
    "streamlit>=1.46.0",
    "transformers>=4.52.4",
]
